Removing the typos from beemovie.txt
The current version of beemovie.txt is, unfortunately, the typo-filled version (with many instances of the letter c being replaced with the letter o, with typos such as 'ooming' instead of 'coming', at position 292). I already had a copy of the bee movie script saved on my computer, which I went through the effort of removing all of these 'c replaced with o' typos from, so I've just removed the non-alphanumeric characters from that and made it lowercase (like the current, incorrect version of beecsv does). And with this, humanity can now have genuine bee seperated values, untarnished by the typographical errors which formerly plagued it.